+++ lib/mobility/plugins/active_record/dirty.rb
@@ -51,12 +51,6 @@ locale suffix, so +title_en+, +title_pt_br+, etc.)
           end
         end
 
-        included_hook do |_, backend_class|
-          if options[:dirty]
-            backend_class.include BackendMethods
-          end
-        end
-
         private
 
         def dirty_handler_methods
@@ -109,8 +103,6 @@ locale suffix, so +title_en+, +title_pt_br+, etc.)
             end
           end
         end
-
-        BackendMethods = ActiveModel::Dirty::BackendMethods
       end
     end